+++ .teamcity/settings.kts
@@ -42,6 +42,9 @@ project {
     buildType(MyPy)
     buildType(UnitTests)
     buildType(Examples)
+    buildType(PipPython310)
+    buildType(PipPython311)
+    buildType(PipPython312)
     buildType(Tests)
 
     template(GitHubIntegrationTemplate)
@@ -49,6 +52,9 @@ project {
     template(MyPyTemplate)
     template(UnitTestsTemplate)
     template(ExamplesTemplate)
+    template(PipPython310Template)
+    template(PipPython311Template)
+    template(PipPython312Template)
 
     features {
         buildTypeCustomChart {
@@ -295,6 +301,90 @@ object MyPy : BuildType({
     }
 })
 
+object PipPython310Template : Template({
+    name = "PipPython310Template"
+
+    detectHangingBuilds = false
+
+    vcs {
+        root(DslContext.settingsRoot, "+:. => imod-python")
+
+        cleanCheckout = true
+    }
+
+    steps {
+        script {
+            name = "Pip install python 3.10"
+            id = "pip_install_py310"
+            workingDir = "imod-python"
+            scriptContent = """
+                    pixi run --environment py310 --frozen test_import
+                """.trimIndent()
+            formatStderrAsError = true
+        }
+    }
+
+    requirements {
+        equals("env.OS", "Windows_NT")
+    }
+})
+
+object PipPython311Template : Template({
+    name = "PipPython311Template"
+
+    detectHangingBuilds = false
+
+    vcs {
+        root(DslContext.settingsRoot, "+:. => imod-python")
+
+        cleanCheckout = true
+    }
+
+    steps {
+        script {
+            name = "Pip install python 3.11"
+            id = "pip_install_py311"
+            workingDir = "imod-python"
+            scriptContent = """
+                    pixi run --environment py311 --frozen test_import
+                """.trimIndent()
+            formatStderrAsError = true
+        }
+    }
+
+    requirements {
+        equals("env.OS", "Windows_NT")
+    }
+})
+
+object PipPython312Template : Template({
+    name = "PipPython312Template"
+
+    detectHangingBuilds = false
+
+    vcs {
+        root(DslContext.settingsRoot, "+:. => imod-python")
+
+        cleanCheckout = true
+    }
+
+    steps {
+        script {
+            name = "Pip install python 3.12"
+            id = "pip_install_py312"
+            workingDir = "imod-python"
+            scriptContent = """
+                    pixi run --environment py312 --frozen test_import
+                """.trimIndent()
+            formatStderrAsError = true
+        }
+    }
+
+    requirements {
+        equals("env.OS", "Windows_NT")
+    }
+})
+
 object UnitTests : BuildType({
     name = "UnitTests"
 
@@ -343,6 +433,24 @@ object Examples : BuildType({
     }
 })
 
+object PipPython310 : BuildType({
+    name = "PipPython310"
+
+    templates(PipPython310Template, GitHubIntegrationTemplate)
+})
+
+object PipPython311 : BuildType({
+    name = "PipPython311"
+
+    templates(PipPython311Template, GitHubIntegrationTemplate)
+})
+
+object PipPython312 : BuildType({
+    name = "PipPython312"
+
+    templates(PipPython312Template, GitHubIntegrationTemplate)
+})
+
 object Tests : BuildType({
     name = "Tests"
 
@@ -377,6 +485,15 @@ object Tests : BuildType({
     }
 
     dependencies {
+        snapshot(PipPython310) {
+            onDependencyFailure = FailureAction.FAIL_TO_START
+        }
+        snapshot(PipPython311) {
+            onDependencyFailure = FailureAction.FAIL_TO_START
+        }
+        snapshot(PipPython312) {
+            onDependencyFailure = FailureAction.FAIL_TO_START
+        }
         snapshot(Examples) {
             onDependencyFailure = FailureAction.FAIL_TO_START
         }
@@ -400,12 +517,33 @@ object Nightly : Project({
     buildType(NightlyUnitTests)
     buildType(NightlyExamples)
     buildType(NightlyTests)
+    buildType(NightlyPipPython310)
+    buildType(NightlyPipPython311)
+    buildType(NightlyPipPython312)
 })
 
 object NightlyLint : BuildType({
     name = "Lint"
 
     templates(LintTemplate)
+})
+
+object NightlyPipPython310 : BuildType({
+    name = "PipPython310"
+
+    templates(PipPython310Template)
+})
+
+object NightlyPipPython311 : BuildType({
+    name = "PipPython311"
+
+    templates(PipPython311Template)
+})
+
+object NightlyPipPython312 : BuildType({
+    name = "PipPython312"
+
+    templates(PipPython312Template)
 })
 
 object NightlyMyPy : BuildType({