+++ imod/mf6/model.py
@@ -4,6 +4,7 @@ import abc
 import collections
 import inspect
 import pathlib
+from collections import defaultdict
 from copy import deepcopy
 from pathlib import Path
 from typing import Optional, Tuple, Union
@@ -25,6 +26,7 @@ from imod.mf6.validation import pkg_errors_to_status_info
 from imod.mf6.write_context import WriteContext
 from imod.schemata import ValidationError
 from imod.typing import GridDataArray
+from imod.typing.grid import ones_like
 
 
 class Modflow6Model(collections.UserDict, abc.ABC):
@@ -486,8 +488,7 @@ class Modflow6Model(collections.UserDict, abc.ABC):
 
         methods = self._get_unique_regridder_types()
         output_domain = self._get_regridding_domain(target_grid, methods)
-        new_model[self._get_diskey()]["idomain"] = output_domain
-        new_model._mask_all_packages(output_domain)
+        new_model.mask_all_packages(output_domain)
         new_model.purge_empty_packages()
         if validate:
             status_info = NestedStatusInfo("Model validation status")
@@ -496,17 +497,28 @@ class Modflow6Model(collections.UserDict, abc.ABC):
                 raise ValidationError("\n" + status_info.to_string())
         return new_model
 
-    def _mask_all_packages(
+    def mask_all_packages(
         self,
-        domain: GridDataArray,
+        mask: GridDataArray,
     ):
         """
         This function applies a mask to all packages in a model. The mask must
-        be presented as an idomain-like integer array that has 0 or negative
-        values in filtered cells and positive values in active cells
+        be presented as an idomain-like integer array that has 0 (inactive) or
+        -1 (vertical passthrough) values in filtered cells and 1 in active
+        cells
+        
+        Parameters
+        ----------
+        mask: xr.DataArray, xu.UgridDataArray of ints
+            idomain-like integer array. 1 sets cells to active, 0 sets cells to inactive, 
+            -1 sets cells to vertical passthrough
         """
+        if any([coord not in ["x", "y", "layer", "mesh2d_nFaces", "dx", "dy"] for coord in mask.coords]):
+            raise ValueError("unexpected coordinate dimension in masking domain")
+
         for pkgname, pkg in self.items():
-            self[pkgname] = pkg.mask(domain)
+            self[pkgname] = pkg.mask(mask)
+        self.purge_empty_packages()
 
     def purge_empty_packages(self, model_name: Optional[str] = "") -> None:
         """
@@ -531,7 +543,7 @@ class Modflow6Model(collections.UserDict, abc.ABC):
     def _get_regridding_domain(
         self,
         target_grid: GridDataArray,
-        methods: dict[RegridderType, str],
+        methods: defaultdict[RegridderType, list[str]],
     ) -> GridDataArray:
         """
         This method computes the output-domain for a regridding operation by regridding idomain with
@@ -542,20 +554,12 @@ class Modflow6Model(collections.UserDict, abc.ABC):
         regridder_collection = RegridderInstancesCollection(
             idomain, target_grid=target_grid
         )
-        included_in_all = None
-        for regriddertype, function in methods.items():
-            regridder = regridder_collection.get_regridder(
-                regriddertype,
-                function,
-            )
+        included_in_all = ones_like(target_grid)
+        regridders =[regridder_collection.get_regridder( regriddertype,function) for regriddertype, functionlist in methods.items() for function in functionlist]
+        for regridder in regridders:
             regridded_idomain = regridder.regrid(idomain)
-            if included_in_all is None:
-                included_in_all = regridded_idomain
-            else:
-                included_in_all = included_in_all.where(regridded_idomain.notnull())
-
-        if included_in_all is None:
-            raise ValueError("No regridder is able to regrid the domain")
+            included_in_all = included_in_all.where(regridded_idomain.notnull())            
+            included_in_all = regridded_idomain.where(regridded_idomain <=0, other = included_in_all)
 
         new_idomain = included_in_all.where(included_in_all.notnull(), other=0)
         new_idomain = new_idomain.astype(int)