+++ imod/typing/grid.py
@@ -269,18 +269,9 @@ def is_spatial_2D(array: xu.UgridDataArray) -> bool:
 def is_spatial_2D(_: object) -> bool:
     return False
 
-
 @typedispatch
 def is_equal(array1: xu.UgridDataArray, array2: xu.UgridDataArray)->bool:
-    # TODO: use .equals function of xugrid once new release has been made
-    if not array1.shape == array2.shape:
-        return False
-    if not (array1.values == array2.values).all():
-        return False
-
-    grid_dataset_1 = array1.ugrid.grid.to_dataset()
-    grid_dataset_2 = array2.ugrid.grid.to_dataset()       
-    return grid_dataset_1.equals(grid_dataset_2)
+    return array1.equals(array2) and array1.ugrid.grid.equals(array2.ugrid.grid)
 
 @typedispatch
 def is_equal(array1: xr.DataArray, array2: xr.DataArray)->bool: